fix sum stuff

+++ src/main/java/org/spongepowered/asm/util/asm/MethodVisitorEx.java
@@ -4,13 +4,13 @@
 
 package org.spongepowered.asm.util.asm;
 
-import org.spongepowered.asm.lib.*;
-import org.spongepowered.asm.util.*;
+import org.spongepowered.asm.util.Bytecode;
+import org.spongepowered.asm.lib.MethodVisitor;
 
 public class MethodVisitorEx extends MethodVisitor
 {
     public MethodVisitorEx(final MethodVisitor mv) {
-        super(327680,  mv);
+        super(327680, mv);
     }
     
     public void visitConstant(final byte constant) {
@@ -18,6 +18,6 @@ public class MethodVisitorEx extends MethodVisitor
             this.visitInsn(Bytecode.CONSTANTS_INT[constant + 1]);
             return;
         }
-        this.visitIntInsn(16,  (int)constant);
+        this.visitIntInsn(16, constant);
     }
 }
